[flow] Make the defense against cyclic tvars more robust 2/2
Summary:
The added test case demonstrates another potential crash. We should have the invariant that when we are in the state of `Forced`, `s.valid` is actually already fully evaluated and forced.  This is not true today, since we set the state to `Forced` after the first encounter of self recursion.

The situation is similar to the following runtime code:

```
const Unforced = 0;
const Forcing = 1;
const Forced = 2;

let state = Unforced;
let memoed = null;

function f() {
  if (state == Forcing) {
    // commenting out the line fixes the issue
    state = Forced;
    // invalid state
    return 1;
  }
  if (state == Forced) {
    return f();
  }
  if (memoed != null) {
    return memoed;
  }
  state = Forcing
  let result = (() => {
      f();
      return f()
  })();
  state = Forced;
  memoed = result;
  return result;
}
```

The fix is to not set the state to Forced after erroring in forcing. Instead, it should only be set at the end.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55769302

fbshipit-source-id: 43c83c50a7137bb3a393ef4077d63b6be42e13ea