[flow] LTI 2.0
Summary:
In this diff, I will introduce a system which I call LTI 2.0™.

## General Idea

In LTI 1.0, we have successfully give most of our users the feeling of LTI: things feel resolved. No more action-at-a-distance (mostly). Most importantly, we can actually freely inspect types and make decisions when dealing with values in `statement.ml`. However, the vision of LTI is still not fully achieved. We still cannot inspect types in arbitrary places, especially in flow_js, even excluding the parts that have to interact with implicit instantiation. That's because we are still using unresolved tvar infra to generate types for annotations, since it can "gracefully" deal with cyclic annotations, either good or bad.

LTI 2.0 is an attempt to change that. We will generate types for annotations in a system similar to type_sig_merge and annotation inference. At every point in time, we don't use unresolved tvars for reference to a type. We use lazy tvar instead. The lazy tvar is always fully-resolved, although it's strictly forbidden to inspect it before we resolve all the elements in a components in env_resolution.

To see this in action, let's look at an example:

```
type Foo = {foo: Bar};
type Bar = {bar: Foo};
```

During init_env, we assign the following type to `Foo` and `Bar` in the environment:

- `T_Foo = OpenT ('Bar', FullyResolved (b))`
- `T_BAR = OpenT ('Foo', FullyResolved (a)`

What exactly is `a` and `b`? It's a lazy Type.t, that will emit an internal error and return any. With this, we can easily generate type for the type alias `Foo` and `Bar` as `{foo: OpenT ('Bar', FullyResolved (b))}` and `{bar: OpenT ('Foo', FullyResolved (a)}`. At the end of component resolution, we finally resolve the type by patching `a` and `b` to point to
`{foo: OpenT ('Bar', FullyResolved (b))}` and `{bar: OpenT ('Foo', FullyResolved (a)}` respectively. The exact implementation of this patching will be covered in a later section.

In some sense, this is like the classic cartoon humor, where you won't fall off the cliff if you don't look down, and if you don't look down long enough, you can actually walk pass the thin air. This is exactly what's happening here. Note that if you force the tvar early, you will indeed get the internal error and any. However, if you force it after component resolution, it will be totally fine, and you can inspect the type in whatever way you like after that.

This example is quite simple, but what about cases when we actually need to do some computation on the type, e.g. `class MyComponent extends React.Component<Props>`. To generate a type for `React.Component<Props>`, we need to do a get-prop and type app. Both operations require inspection of the input type. The key to resolve this is to lazily map the lazy fully resolved OpenT. If we obey the constraint that we do not early force it, then in the lazy map, we can inspect the type, since at runtime, the inspection and forcing only happens after component resolution.

## Patching Environment Tvars

In this implementation, the lazy type in the environment tvars doesn't just store a type. Instead, it will simply deref a `Type.t lazy_t ref` stored in the loc_env along side with the environment tvar. Initially, this mutable ref contains a lazy type that will emit an internal error and return any. Later when we resolve it, we will patch this ref with a lazy type that contains the final result of resolution.

This system completely replaces the previous system where we store a set of env entries that are readable or under resolution.

## Lazy Mapping Complications

It sounds easy to say that to do some lazy operation on lazy type, we just do `Lazy.map ~f:ops lazy_t`. We did already do that to some extent in type_sig_merge and annotation inference. However, the system with flow_js is much more complicated that brings extra complication.

One issue that annotation inference doesn't have to deal with is that annotation inference never runs in a speculation context, while this is possible in statement.ml + flow_js. Speculation will throw to indicate that a branch has failed. It will be catastrophic if we throw in the lazy block, because OCaml's lazy will remember it, and rethrow the same exception every time we force the lazy...

Fortunately, it doesn't quite make sense for these annotations' checking to run in speculation context anyways. It's mostly a side effect of problematic design of EvalT. In principle, the evaluation of the annotations should never determine which branch we take in speculation. If the type is bad, like `Prop['does_not_exist']`, then it should just be any, and error unconditionally. Therefore, I disabled speculation in these lazy tvar mappings (in `Tvar.map_fully_resolved`).

Another issue is that we need to make sure we are not storing OpenT in the payload of another fully resolved OpenT. This is achieved by unwrapping the type. I used the unwrapping in type_hint that will unwrap both `AnnotT` and `OpenT`. You might wonder why we also unwrap `AnnotT`. It turns out that's the key to detect cyclic types!

Finally, these mappings can create cyclic types that cannot be forced. That's why we need D54021755 to defend against these cases.

## Forcing the Lazy Types

The LTI 2.0 system as described above, has an effect of being too lazy, which deviates from the mostly eager behavior we have right now. If we allow it to happen, then we might miss some errors like `value-as-type` that we currently raise.

To restore the current behavior, we will force all the lazy tvars that is generated within a component, and post inference. This is achieved by registering such lazy tvar during its creation time in the context, and forcing all of them at the end. Some extra care needs to be taken for environment tvars, since they are all created during init_env. We will skip the registration for these tvars, but remember to resolve them as well at the end of component resolution in env_resolution.

Currently, some EvalTs are evaluated immediately. This will break the invariant that types cannot be forced early. To address this, I used the same infra to encode the force evaluation of EvalT in a lazy tvar, so that it will only be forced at the end of component resolution.

The eager forcing + unwrapping AnnotT + builtin defense against cyclic tvar in the previous diff bring us a nice benefit: we can finally detect cyclic annotations!

Changelog: [errors] We rewrite the way we generate types for annotations. We will now detect and error on trivially recursive types like `type T = T` or `type Foo = typeof foo; const foo: Foo = ...`. In addition to this, you might see some errors being moved around.

Reviewed By: panagosg7

Differential Revision: D53460591

fbshipit-source-id: 24705a5d61a909cac26ed64bec3ba5ea1c430b94