[flow][refactor] Move `Type_operation_utils.TvarUtils` to `flow_js_utils`
Summary:
For earlier iterations of these utils, it used to depend on flow_js and also uses `Flow_js_utils.merge_tvar`, so it also accepts `filter_empty`. In the landed LTI 2.0 diff, none of these are true, so let's clean it up and move to `flow_js_utils`, so it can be called in flow_js as well.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56023911

fbshipit-source-id: 80b41a724ec3ca8d48da283e2689214c62db2bc5