[flow][autocomplete] simplify keyword search
Summary:
In follow up diffs, we will be making some small but subtle adjustments to the ASTs that we pass into autocompleteService_js.ml.

When the input is `foo.bar|` we will be first converting this to a canonical form `foo.AUTO332`. We will then parse this as a canonical AST that we will run inference on. Finally, we will "patch" this AST so that identifiers with name `"AUTO332"` are replaced with `"barAUTO332"`, before we run the autocomplete analysis.

The issue with the previous version of the keyword mapper was an optimization it did that would not allow it to descend into subtrees that did not contain the target `"barAUTO332"` entirely. Since the patching we do only changes the identifier node, and not its parents, this optimization would cause us to miss `"barAUTO332"` due to the locations in the parent nodes of that identifier not matching the target.

Given that trying to "patch" all locations in parent nodes in the AST is a bit tedious, this diff takes a different approach. It makes this search robust to the upcoming plan, by only considering as target the first character of the identifier that we run autocomplete on. This location is invariant to the canonicalization that we do, so the results of the keyword search are the same.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55409886

fbshipit-source-id: a7f21d6a8279773840f682ac10e0c3c178621bc8