[flow] Kill ConditionalTypeEval use_op
Summary:
I re-discovered this while looking through the error printing code, and found the message for that use_op not that meaningful.

This use_op should never appear to the end user, since the evaluation of conditional type should never fail (especially after the previous diff to ensure that we no longer error on underconstrained infer).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56041988

fbshipit-source-id: 08aa0b80abf94bdb1ca1df66b29883b02d05908c