[flow] move matching prop test to statement.ml
Summary:
The matching-prop check establishes that in tests of the form
```
if (obj.prop === 'abc') { ... }
```
`'abc'` is one one potential types of the `prop` property of `obj`.

Until now, this check was set up in the name resolver and was limited in the form of expressions of `obj` that were allowed. Specifically `obj` needed to be an identifier. So the check would not be performed on tests of form
```
if (obj.propA.propB === 'abc') { ... }
```
We want to lift these restrictions and fire this test more reliably and predictably. To do so we are moving this check to statement.ml. After all this test is not relevant to the information that the Name_resolved is concerned with.

For now, we are focusing on preserving the semantics of this check: we still only perform the check on code of the form of the first example. To achieve this we make `of_expression` available from statement.ml. The next diff generalizes.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D56006607

fbshipit-source-id: 266a0198c4ee2c4c25f0376e3617ce1f3800d568