[flow][refactor] Intermediate representation of errors
Summary:
As I want to improve error messages, I want to first establish a more solid foundation. Currently, the majority of the error message printing logic is done in `flow_error.ml` (a lot of variants are done in elsewhere, but the most interesting subtyping related ones are all in `flow_error`). Currently, `flow_error` is doing too many things in a single pass:

1. Determine error location
2. Convert `Error_message.t` into a form of (root, message, frame, explanation)
3. Generating most of the text and ref of errors, with only the platform-specific formatting deferred to later stage.

Having all of the logic in one pass making is very hard to reason about what's going on, and also makes it hard to optimize it. e.g. for error collation, we still need to generate the friendly error message, even if we only care about error location.

In this diff, I will start to tame this mess. We will split 1,2,3 into two steps (1,2) and 3. Since 1 and 2 are highly related, they are still done in a single pass, where we go from `Error_message.t` into an error IR. The error IR mostly uses variants that keeps only necessary information to compute the printed message. Then in a second step, we convert the error IR into printable error.

This diff only does the refactor. We are not taking advantage of any of the potential perf opportunities yet. I also choose not to clean up some of the inconsistencies I observed during this refactor, so that we can have a clean error diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56087851

fbshipit-source-id: c592b429f4363b5287757a6e5afe0a72252b5c6d