[flow][errors] Eliminate last usage of `ExplanationAlreadyPrinted`
Summary:
I want to ensure that `intermediate_error` contains no `Friendly.message_feature`, which contains a lot of strings that are mostly constant, but very wasteful to send between processes. The end goal is that we will send `intermediate_error` to the main process instead, which will contain already decided primary error location.

This diff starts with the easy one: we get rid of `ExplanationAlreadyPrinted`. We just need to add two more variants.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56164157

fbshipit-source-id: 4e73c5a1a9ea7ffc97288b13b162d4911956f34d