[flow] Do not generate ExactT in flow_js_utils
Summary:
The larger goal is to replace the `ExactT` implementation with a `EvalT` based one, which is more battle-tested and has better handling for things like distribute-over-union.

This part is factored out of D56071682 so that we can start replacing some `ExactT` when it's not necessary. But it turns out that this part can also be landed without issues.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D56279214

fbshipit-source-id: 06903c0c96eca8f4fb4e14742e6c8a97188af5cb