[flow][cleanup] Turn EPredicateFuncTooShort into an internal error
Summary:
The error message mentions $Refine, which is killed last year and no longer possible to construct. Therefore, the codepath that generates the error shouldn't ever be reached. I noticed it when doing error message refactors.

This diff replaces it with an internal error instead, since it would never be shown to the user.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56272907

fbshipit-source-id: c7928af793b0c10bb22a15cac718612a4d4b97e2