[flow][refactor] Move most of the error printing logic in error_message into flow_intermediate_error
Summary:
I want to ensure that `intermediate_error` contains no `Friendly.message_feature`, which contains a lot of strings that are mostly constant, but very wasteful to send between processes. The end goal is that we will send `intermediate_error` to the main process instead, which will contain already decided primary error location.

This diff does most of the work to move the error printing logic in error_message into flow_intermediate_error. We create new `'loc message` variants, and printing them in flow_intermediate_error. We only do a light transform in `error_message.ml`. In the future, most of the `EXyz` can be replaced with `ESimple of {loc; code; msg}`

This diff doesn't completely finish the migration. There are still a few of them that I either suspect to be unused, or just haven't got to them yet.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56172516

fbshipit-source-id: 30878d91dcb02d52645b83d5f82ca9abdbed4caf