[flow][refactor] Store `Flow_intermediate_error_types.message` in `UseOp`'s attached message instead of `Friendly.message_feature`
Summary:
I want to ensure that `intermediate_error` contains no `Friendly.message_feature`, which contains a lot of strings that are mostly constant, but very wasteful to send between processes. The end goal is that we will send `intermediate_error` to the main process instead, which will contain already decided primary error location.

This is the last place we store  `Friendly.message_feature`. This diff first makes it store `Flow_intermediate_error_types.message`, so that all of them can be later migrated to variants, instead of this structure with a lot of strings.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56317082

fbshipit-source-id: 5e1cd7e58c8d1cdf6fa9530a321563d09a49bfd7