[flow][refactor] Migrate all remaining `MessageAlreadyFriendlyPrinted` to variants
Summary:
I want to ensure that `intermediate_error` contains no `Friendly.message_feature`, which contains a lot of strings that are mostly constant, but very wasteful to send between processes. The end goal is that we will send `intermediate_error` to the main process instead, which will contain already decided primary error location.

This diff finishes the migration, by creating a variant for every message in UseOp recipe.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56328507

fbshipit-source-id: d44b82e762014e6babdb09e82ee2db494255cf80