[flow][config][EZ] Introduce flag to control faster error collation
Summary:
Introduce a flag that will gate faster error collation which I will introduce behind a flag. I believe it will be very safe, but gate it so that we can easily measure perf improvements, isolated from other improvements and regression that might be in the same release.

For now, I set it to false, so that we can easily see the next diff has almost identity perf characteristics with the current implementation. It will be flipped to true in the next diff after the next diff.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D56378526

fbshipit-source-id: 65a3974e72c5fa52d3793664bd46ec0392476650