[flow] Skip error printing for suppressed errors under `faster_error_collation=true`
Summary:
This is the diff that starts to take advantage of the hard work to create the intermediate representation of error without any printing logic.

Currently, we do a lot of wasted work of converting `Error_message.t` into printable errors (I will just call it printing below). This can be expensive, since for every reason description, we will scan the string to turn all the stuff wrapped in backticks into `Code` form. We do this even for suppressed errors, since before the error IR change, the logic of determine primary location of error is hopelessly entangled together with printing logic.

Now with error IR, this is no longer necessary, we just need to do a relatively straightforward transformation of error that's mostly constructing objects and occasionally to a `loc_of_aloc` conversion to do some location comparison. To determine whether an error is suppressed, we just need to perform `Error_message.t -> intermediate_error` transform without printing, so we can skip the step of printing all the errors that end up being suppressed anyways (which is most of the errors).

Note that this is not the end of the optimizations unlocked by error IR yet. We still haven't moved the `Error_message.t -> intermediate_error` transform onto the worker processes, which can relieve the stress on the main process even more. But I will stop here in this diff for now, since I want to gate it to measure perf win.

For the gating, I just forked the logic in error_suppressions and error_collator, by copying the old code and suffix them with `_old`. They should be killed soon once we have the perf number.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56332422

fbshipit-source-id: a2d3010356e0183d08ccf807d3a20a503e53e24f