[flow][easy] Remove unused deprecation alerts
Summary:
In OCaml 5, there is a new "unused attribute" warning which fires for any unused
attributes. This is a handy warning and it shows that these two alerts are
unused.

I did a bit of digging, and it seems that OCaml does not support alerts on this
position at all, neither in the implementation nor the interface.

Reviewed By: SamChou19815

Differential Revision: D56434654

fbshipit-source-id: be45812e643d6c0af735708316e75c0ed07a6446