[flow][autocomplete] fix locs for all calls to Autocomplete_sigil.remove_from_loc
Summary:
The function `Autocomplete_sigil.remove_from_loc` in classic mode assumes tokens of the form `fooAUTO332bar`, and returns the loc span of the token with the `AUTO332` part removed.

In canonical mode, the respective tokens will be of the form `AUTO332`. Before this diff, we would adjust the location we pass into this function in `autocomplete_get_results`, and so did not have to account for this discrepancy in `remove_from_loc`. We did not perform the same adjustment in other calls to `remove_from_loc`, which led to a small discrepancy in the full output of autocomplete.

An example of the discrepancy was in the edit location of the `[Symbol.iterator]` entry in the following request (see also example that changes):
```
function takesMap(map: Map<number, string>) {
    map.si
//        ^
}
```

This diff addresses this by making `remove_from_loc` add the length of the prefix and suffix of the relative token to the result. So, if the relative token is `fooAUTO332bar` it will add `len("foo") + len("bar") = 6` to the location ending column.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D56448499

fbshipit-source-id: 718020e84cbddf91d323c27d4dc646b51950bea8