[flow] Eliminate construction of `RCustom` in flow_intermediate_error through hardcoded_string_desc_ref
Summary:
The goal is to eliminate `RCustom`, which has no structure whatsever, and prevents us from making improvements to errors.

It's currently constructed very commonly in error printing. However, it actually has very little to do with reason, but we just use it to construct a reference pointers with hardcoded inline messages.

It's debatable whether we should allow these kinds of arbitrary strings in error generation, but at least with the error IR work, it matters much less, so I don't think it's worth it to convert them into variants as well. However, it's still important to get rid of RCustom. Therefore, I introduced `hardcoded_string_desc_ref` that will allow us to directly creates such refs in error, so that we don't have to achieve it indirectly via RCustom.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56492702

fbshipit-source-id: 4d2e3d1d84a8e06b68b2a70280a90c8acbc28163