[flow][ast] Use "kind" field for type guards
Summary:
We change the way we represent the kind of a type guard to make it easier to add more variations (in upcoming diffs).

For now the two variations we have are the default case (`x is T`) and the assertion type guard (`asserts x is T`). The latter is not supported in typing but it is parsed as expected. The new `kind` field takes the values `Default` and `Asserts` respectively in the Flow AST. In estree, a new "kind" field is added, that replaces the "asserts" field, and takes values `null`, and `"asserts"` respectively.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D56553374

fbshipit-source-id: add25157e2e108ead12351ae20cb3a58ed511dc6