[flow][refactor] Directly grab reason from types for Error_message.EAnnotationInference errors
Summary:
I want to reduce our reliance on reason_of_t, since it also grabs the reason from OpenT, which might be different from the reason of the underlying type. This diff cleans up some callsites when it's completely unnecessary.

In these cases modified in the diff, we can just grab the reason as we pattern match on the types.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56642931

fbshipit-source-id: d53df66c235965c5e8089de2cfd99bb182ffe9e3