[flow] Replace `RCustom "providers"` with an identifier description
Summary:
We should never show providers as the reason to the user. Instead, in all the call sites, we have access to the identifier name, so we should use that as reason description instead.

In these cases, it doesn't quite matter, since the reason is only accessible from typed ast, but it's still better than RCustom.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D56714487

fbshipit-source-id: 5496073588df5c08dd21d4d20b7061a58c563abc