[flow] Kill support for traces (user-visible part)
Summary:
The traces feature is not that useful nowadays given that we have use_op that makes the error much more understandable. Althrough our errors right now is not perfect, what's currently in the traces are completely non-comprehensible. Instead of dragging on, let's just delete this feature.

This diff starts from the user visible part. We will stop tracking it in errors, and remove the config options. `Context.max_trace_depth` is hardcoded to 0 for now (value when the flag is off), which will be cleaned up in a later diff.

Changelog: [breaking] Support for `--traces` and the corresponding flowconfig has been removed.

Reviewed By: panagosg7

Differential Revision: D56735978

fbshipit-source-id: 0baf317b39ed5893724c3a1118b2fdfdf2fb3afb