[flow][cleanup] Cleanup `Context.max_trace_depth` and simplify trace
Summary:
With some of the bigger mechanical cleanup done, it's time to simplify the structure of trace. Since now we never inspect the type part of trace, the only thing recorded in Type.trace that's relevant is the depth. This diff does the simplification.

Note that this part cannot be killed, since the recursion limit check depends on this, so we still need to pass traces around everywhere. In the future, we can expand on the trace param to make it a env param for flow_js, which can keep track of some flags like `in_implicit_instantiation` that should be moved out of Context.t.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56735979

fbshipit-source-id: f9450bde33742e30a7c0881a275dcd810fbf6bab