[flow] add shortcut for ElemT dict access
Summary:
This is the outcome of a recent performance investigation. The problematic scenario was the following
```
declare var x: LargeUnion;
declare var dict: { [key: LargeUnion]: V };
const v = dict[x];
```
The generated constraint is the following
```
l ~> ElemT (ObjT ({ [key: k]: v }, tout) with l = k = LargeUnion
```
Before we would not recognize that the dictionary key `k` had the same type as the accessing value `l`. This would cause the large union on the left to be broken up into its parts leading to slowdowns.

This diff addresses this scenario by performing a lightweight check on `k` and `l`. If they're the same, we short-circuit and return `v` to `tout`.

We make sure to propagate the deep-readonly-ness if the flag exists on the input object.

Changelog: [perf] performance improvement of checking dictionary accesses with keys that are equivalent to the dictionary key.

Reviewed By: SamChou19815

Differential Revision: D56047143

fbshipit-source-id: 46fcccd871db8e27c116dec4a619a6b1f4970214