[flow] Builtin await typing through conditional type infra instead of relying on $await hack
Summary:
In this diff, we replace the two sites where to do the hacky `$await -> CallT` to unwrap Promise with a builtin implementation, powered by conditional type infra.

Since we are dealing with expressions, we have LTI guarantee, so we can inspect the input type. Therefore, we can easily implement the distribute-over-union behavior that currently exist for the `await` operator. As a side effect, we can remove a bunch of spurious underconstrained errors, since the EmptyT case is handled automatically, instead of hitting the case where `EmptyT ~> Promise<T>` where we miss empty propagation.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D56861865

fbshipit-source-id: 431d3a76d42ae4159cfe5914266bd76d52a656e7