[flow] fix printing of tuple elements with variance
Summary:
The type `$ReadOnly<[a: number, string]>`, can also be expressed by passing the polarity to the tuple element. To do so, however, we need to have a label for each element. (The form `[+number]` is not parseable.)

We change print to output a made-up name `_i` (where `i` the order of the element) for elements with non-neutral variance. E.g. `[+a: number, +_1: string]`.

Same holds for optional element.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D56913842

fbshipit-source-id: 878b95fcac33588df98344811fc8dba2371a0748