[flow] Remove last mention of $CharSet in libdef
Summary:
In this series of diffs, I'm going to kill $CharSet support. It's only used in one place in our libdef (deleted in this diff), and only used in flow-typed that is copy of ours: https://github.com/search?q=repo%3Aflow-typed%2Fflow-typed+%24CharSet&type=code

It's probably better to just use string. If users really care, it can be enforced by a linter.

Changelog: [internal] grouped together with the last diff in the stack

Reviewed By: gkz

Differential Revision: D56716721

fbshipit-source-id: a49dd3c6cd94f23a5e54387f366d4e3ac74c225a