[flow] Ignore the content of `exact_reason` in `Slice_utils.mk_object`
Summary:
As the effort to replace `ExactT` with EvalT-based implementation, I want to lock down all places where we construct `ExactT`. This is the final non-EvalT-like use case, and we still haven't completely figured out why it's necessary to preserve existing behavior and perf.

This diff aims to show that at least reason is not the reason this is critical, since we can ignore the reason content, since the last call site passes in a reason for `exact_reason` to be the same as the `def_reason`.

Building on this knowledge in the next diff, I will rename this param to avoid future confusion.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D56916901

fbshipit-source-id: 2c9002247282b50614962a17c42c2b460e7810b1