[flow][name-resolver] make Object.freeze({...}) a write
Summary:
The name-resolver did not consider the expression `Object.freeze({f: 1})` as a write, but it does with it's argument `{f: 1}`. This leads to the env_resolver calling into `Statement.object_` on `{f: 1}` with `frozen=false`.

This result is unnecessary, since we should only be visiting the object argument to `freeze()` with `froze=true`. This is actually what happens when we finally visit the AST node for the entire expression.

So, instead of visiting the object twice with different values for `frozen` we can visit it once as part of the entire `freeze()` call. The result will be cached and so we won't be needing to visit it again.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D56962883

fbshipit-source-id: 7036eafcbfabb134e9a8bd76687671494c19e339