[flow] Type `React.cloneElement` in the userland
Summary:
I want to make some changes to `React$CreateElement`, and it's very annoying that it shares quite a lot of code with `React$CloneElement` that's also builtin.

Now that Flow supports conditional types, in principle we have the full expressive power to move both typing into userland. However, `createElement` is on the critical path of implicit instantiation, and we also want to provide better errors on the much more common case of jsx construction, it's not worth changing  `React$CreateElement`, so I just moved `React$CloneElement` in the diff.

Most of the complexity of the change is due to the very special behavior of rest children params. If only one static children is passed, then that child is passed as children as is, instead of wrapping it in an array. I was able to detect it using the combination of overloads and conditional type. It certainly regresses the error quality a bit, but people really shouldn't use this API.

Changelog: [errors] We changed the typing of `React.cloneElement`. You might see errors being shifted around, and some advanced patterns no longer supported. [The API is not encouraged by React](https://react.dev/reference/react/cloneElement), so migrating away from it is recommended.

Reviewed By: panagosg7

Differential Revision: D56551254

fbshipit-source-id: 39d27a8ae7c59d5e5b3f5b4f244de7c90b0929db