[flow] Flag for detailed diagnostics for vscode
Summary:
This is the feature flag intended for a feature similar to rust-analyzer's "Click to see full compiler diagnostics" feature.

I specially put VSCode in the feature flag name. In principle, extensions for other editors could speak the same custom protocol, we simply do not have enough bandwidth to support them, especially if we want to change the rendering protocol.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D57169090

fbshipit-source-id: 300ff15e97499e0843c76f03fe35bf13d88d61bc