[flow] Cleanup unused tty style with background
Summary:
Later in the stack, I want to propagate these styling descriptor through the LSP to power the full detailed error rendering in vscode, and the same also needs to be implemented on the vscode side. The fewer style we support, the easier it will be for us to maintain.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D57594525

fbshipit-source-id: 0f373c9668bc74ace7eafc5cb0df72270adba256