[flow][detailed-error] Colorized errors with clickable links in IDE
Summary:
Right now for CLI, we have two modes:

1. Colored output with coalesced references: closer references are grouped into a single block. This mode requires color, because otherwise user will be unable to distinguish which reference points to in the inline code block.
2. Colorless output without coalesced references

For VSCode detailed errors, none of these are a perfect fit. We certainly want colors, but the coalesced reference part is not ideal:

1. We have an entire editor screen to render it, so some repetition doesn't quite matter.
2. With coalesced references, locations for each reference are no longer printed together, especially the line number are not printed. This is not a good experience in IDE when user wants to directly click to go to a reference e.g.

```
test.js
[1] 42 | const a: string = <colored>3</colored>
```

This diff implements the behavior of the new IDE coloring mode: we use the plaintext printing as the base, since it will print out the full location of every reference. Then we will apply the color to all references using the same printing strategy we use for colored printing mode:

1. dim for code block
2. rotating colors for each reference

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D57848219

fbshipit-source-id: fdbfb6b42f94ab0b3384704b0a6399726199be27