[flow] do not wrap TypeTs in hooklike
Summary:
An import of a type that had a hooklike name used to be wrapped in an `EvalT MakeHookLike`. This caused issues with normalization since the normalizer did not expect to see an `EvalT` as the toplevel of a type alias.

Hooklike should not have any special meaning for type aliases, so this diff fixes this issue by not applying the `make_hook_like` wrapping on `TypeT`s.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D57933461

fbshipit-source-id: a632ee6de1333b054ccf7624f39e9c9487137641