[flow] show specialized type on hover over polymorphic JSX call to component (new syntax)
Summary:
Before, hovering type on
```
declare component Foo<T>(x: T, y: T);
<Foo x={1} y={""} />;
// ^
```
would return `compont Foo<T>` , instead of the specialized form based on the inferred type argument to the element creation.

With this diff Flow will now show `component Foo<number | string>`.

Notice that this form is not a proper type. In ty.ml it is still represented as a `NominalComponentDecl` with an extra field for type arguments, instead of just type parameters. When type arguments are available, we show them instead of the tparams.

Also note that since the syntax `component C<Type>` is toplevel only, when we have a toplevel union or intersection of components, we fall back to `React$AbstractComponent<...>`. This case should be rare. (See added examples.)

The implementation reuses the "specialized callee" machinery that we use for polymorphic function calls (D47180573).

One slight deviation from that approach is that we call `CalleeRecorder.add_callee` to record the instantiated version of the component in react_kit.ml instead of flow_js.ml. We store the inferred targ list within the `React.CreateElement` constructor and use it in `React_kit.create_element` when the component type gets concretized.

Changelog: [fix] hover-types will now show the specialized version of a polymorphic component used at a JSX expression

Reviewed By: SamChou19815

Differential Revision: D57764782

fbshipit-source-id: 3cdf130a75018ef7f66f3844f70d74675f26ddbd